requirements.txt
flask
requests
beautifulsoup4
pandas
openpyxl
pillow